refactor: polish
