+++ src/commands/new/parachain.rs
@@ -1,134 +1,27 @@
 use crate::{
-	engines::parachain_engine::{instantiate_template_dir, Config},
-	git::{GitHub, TagInfo},
-	helpers::git_init,
+	engines::{
+		parachain_engine::{instantiate_template_dir, Config},
+		templates::{Provider, Template},
+	},
+	git::GitHub,
+	helpers::{display_release_versions_to_user, git_init},
 	style::{style, Theme},
 };
 use anyhow::Result;
-use clap::{Args, Parser};
+use clap::Args;
 use std::{fs, path::Path};
-use strum_macros::{Display, EnumString};
-use url::Url;
 
 use cliclack::{clear_screen, confirm, input, intro, outro, outro_cancel, set_theme};
-
-#[derive(Clone, Parser, Debug, Display, EnumString, PartialEq)]
-pub enum Template {
-	#[strum(serialize = "Pop Base Parachain Template", serialize = "base")]
-	Base,
-	#[strum(serialize = "OpenZeppeling Runtime Parachain Template", serialize = "template")]
-	OZTemplate,
-	#[strum(serialize = "Parity Contracts Node Template", serialize = "cpt")]
-	ParityContracts,
-	#[strum(serialize = "Parity Frontier Parachain Template", serialize = "fpt")]
-	ParityFPT,
-}
-impl Template {
-	fn is_provider_correct(&self, provider: &Provider) -> bool {
-		match provider {
-			Provider::Pop => return self == &Template::Base,
-			Provider::OpenZeppelin => return self == &Template::OZTemplate,
-			Provider::Parity => {
-				return self == &Template::ParityContracts || self == &Template::ParityFPT
-			},
-		}
-	}
-	fn from(provider_name: &str) -> Self {
-		match provider_name {
-			"base" => Template::Base,
-			"template" => Template::OZTemplate,
-			"cpt" => Template::ParityContracts,
-			"fpt" => Template::ParityFPT,
-			_ => Template::Base,
-		}
-	}
-	fn repository(&self) -> Url {
-		match &self {
-			Template::Base => Url::parse("https://github.com/r0gue-io/base-parachain")
-				.expect("valid pop base template repository url"),
-			Template::OZTemplate => {
-				Url::parse("https://github.com/OpenZeppelin/polkadot-runtime-template")
-					.expect("valid openzeppelin template repository url")
-			},
-			Template::ParityContracts => {
-				Url::parse("https://github.com/paritytech/substrate-contracts-node")
-					.expect("valid parity substrate contract repository url")
-			},
-			Template::ParityFPT => {
-				Url::parse("https://github.com/paritytech/frontier-parachain-template")
-					.expect("valid partity frontier repository url")
-			},
-		}
-	}
-}
-
-#[derive(Clone, Default, Parser, Debug, Display, EnumString, PartialEq)]
-pub enum Provider {
-	#[default]
-	#[strum(serialize = "Pop", serialize = "pop")]
-	Pop,
-	#[strum(serialize = "OpenZeppelin", serialize = "openzeppelin")]
-	OpenZeppelin,
-	#[strum(serialize = "Parity", serialize = "parity")]
-	Parity,
-}
-impl Provider {
-	fn default_template(&self) -> Template {
-		match &self {
-			Provider::Pop => return Template::Base,
-			Provider::OpenZeppelin => return Template::OZTemplate,
-			Provider::Parity => return Template::ParityContracts,
-		}
-	}
-	fn from(provider_name: &str) -> Self {
-		match provider_name {
-			"Pop" => Provider::Pop,
-			"OpenZeppelin" => Provider::OpenZeppelin,
-			"Parity" => Provider::Parity,
-			_ => Provider::Pop,
-		}
-	}
-	fn display_select_options(&self) -> &str {
-		match &self {
-			Provider::Pop => {
-				return cliclack::select(format!("Select a template provider: "))
-					.initial_value("base")
-					.item("base", "Base Parachain", "A standard parachain")
-					.interact()
-					.expect("Error parsing user input");
-			},
-			Provider::OpenZeppelin => {
-				return cliclack::select(format!("Select a template provider: "))
-					.initial_value("template")
-					.item(
-						"template",
-						"OpenZeppeling Template",
-						"OpenZeppeling Runtime Parachain Template",
-					)
-					.interact()
-					.expect("Error parsing user input");
-			},
-			Provider::Parity => {
-				return cliclack::select(format!("Select a template provider: "))
-					.initial_value("cpt")
-					.item("cpt", "Parity Contracts", "A parachain supporting WebAssembly smart contracts such as ink!.")
-					.item("fpt", "Parity EVM", "A parachain supporting smart contracts targeting the Ethereum Virtual Machine (EVM).")
-					.interact()
-					.expect("Error parsing user input");
-			},
-		};
-	}
-}
 
 #[derive(Args)]
 pub struct NewParachainCommand {
 	#[arg(help = "Name of the project. Also works as a directory path for your project")]
 	pub(crate) name: Option<String>,
-	#[arg(help = "Provider to pick template: Options are pop, openzeppelin and parity.")]
+	#[arg(help = "Template provider among 'pop', 'openzeppelin' and 'parity'.")]
 	#[arg(default_value = "pop")]
 	pub(crate) provider: Option<Provider>,
 	#[arg(
-		help = "Template to use. Options are base for Pop, template for OpenZeppelin and cpt and fpt for Parity templates"
+		help = "Template to use. Options are 'base' for Pop, 'template' for OpenZeppelin and 'cpt' and 'fpt' for Parity templates"
 	)]
 	pub(crate) template: Option<Template>,
 	#[arg(long, short, help = "Token Symbol", default_value = "UNIT")]
@@ -147,9 +40,10 @@ pub struct NewParachainCommand {
 impl NewParachainCommand {
 	pub(crate) async fn execute(&self) -> Result<()> {
 		clear_screen()?;
+		set_theme(Theme);
 		if let Some(name_template) = &self.name {
-			let provider = &self.provider.clone().unwrap_or(Provider::Pop);
-			let template = &self.template.clone().unwrap_or(provider.default_template());
+			let provider = &self.provider.clone().unwrap_or(Provider::Pop); //Provider by default POP
+			let template = &self.template.clone().unwrap_or(provider.default_template()); //Each provider has a template by default
 			if !template.is_provider_correct(provider) {
 				outro_cancel(format!(
 					"The provider \"{}\" doesn't support the {} template.",
@@ -164,15 +58,16 @@ impl NewParachainCommand {
 			};
 			generate_template(name_template, provider, template, None, config)?
 		} else {
-			guide_user().await?;
+			guide_user_to_generate_template().await?;
 		}
 		Ok(())
 	}
 }
 
-async fn guide_user() -> Result<()> {
+async fn guide_user_to_generate_template() -> Result<()> {
+	intro("User guide to generate a template")?;
 	let name: String = input("Where should your project be created?")
-		.placeholder("my-parachain")
+		.placeholder("./my-parachain")
 		.interact()?;
 
 	let provider_name = cliclack::select(format!("Select a template provider: "))
@@ -190,17 +85,19 @@ async fn guide_user() -> Result<()> {
 	let template_name = provider.display_select_options();
 	let template = Template::from(template_name);
 
-	// Get the releases
-	let url = template.repository();
+	let url = url::Url::parse(template.repository_url()).expect("valid repository url");
 	let latest_3_releases = GitHub::get_latest_releases(3, &url).await?;
 
-	let tag_name = display_versions(latest_3_releases)?;
+	let mut release_name = None;
+	if latest_3_releases.len() > 0 {
+		release_name = Some(display_release_versions_to_user(latest_3_releases)?);
+	}
 
 	generate_template(
 		&name,
 		&provider,
 		&template,
-		Some(tag_name),
+		release_name,
 		Config {
 			symbol: "UNIT".to_string(),
 			decimals: 12,
@@ -223,7 +120,6 @@ fn generate_template(
 		template,
 		provider
 	))?;
-	set_theme(Theme);
 	let destination_path = Path::new(name_template);
 	if destination_path.exists() {
 		if !confirm(format!(
@@ -251,54 +147,6 @@ fn generate_template(
 	spinner.stop("Generation complete");
 	outro(format!("cd into \"{}\" and enjoy hacking! 🚀", name_template))?;
 	Ok(())
-}
-
-fn display_versions(latest_3_releases: Vec<TagInfo>) -> Result<String> {
-	let version;
-	if latest_3_releases.len() == 3 {
-		version = cliclack::select(format!("Select a specific release:"))
-			.initial_value(&latest_3_releases[0].tag_name)
-			.item(
-				&latest_3_releases[0].tag_name,
-				&latest_3_releases[0].name,
-				format!("{} ({})", &latest_3_releases[0].tag_name, &latest_3_releases[0].id),
-			)
-			.item(
-				&latest_3_releases[1].tag_name,
-				&latest_3_releases[1].name,
-				format!("{} ({})", &latest_3_releases[1].tag_name, &latest_3_releases[1].id),
-			)
-			.item(
-				&latest_3_releases[2].tag_name,
-				&latest_3_releases[2].name,
-				format!("{} ({})", &latest_3_releases[2].tag_name, &latest_3_releases[2].id),
-			)
-			.interact()?;
-	} else if latest_3_releases.len() == 2 {
-		version = cliclack::select(format!("Select a specific release:"))
-			.initial_value(&latest_3_releases[0].tag_name)
-			.item(
-				&latest_3_releases[0].tag_name,
-				&latest_3_releases[0].name,
-				format!("{} ({})", &latest_3_releases[0].tag_name, &latest_3_releases[0].id),
-			)
-			.item(
-				&latest_3_releases[1].tag_name,
-				&latest_3_releases[1].name,
-				format!("{} ({})", &latest_3_releases[1].tag_name, &latest_3_releases[1].id),
-			)
-			.interact()?;
-	} else {
-		version = cliclack::select(format!("Select a specific release:"))
-			.initial_value(&latest_3_releases[0].tag_name)
-			.item(
-				&latest_3_releases[0].tag_name,
-				&latest_3_releases[0].name,
-				format!("{} ({})", &latest_3_releases[0].tag_name, &latest_3_releases[0].id),
-			)
-			.interact()?;
-	}
-	Ok(version.to_string())
 }
 
 #[cfg(test)]
